feat: move ribbon from polygon to edge
