+++ __tests__/unit/mark/geometry/edge.spec.ts
@@ -24,7 +24,7 @@ describe('Edge', () => {
         { type: 'maybeTitleX' },
         { type: 'maybeTooltipY' },
       ],
-      shapes: ['edge', 'smoothEdge'],
+      shapes: ['edge', 'arc', 'smoothEdge', 'ribbon'],
     });
   });
