+++ __tests__/unit/runtime/polygon.spec.ts
@@ -1,5 +1,4 @@
 import { G2Spec, render } from '../../../src';
-import { SANKEY_DATA } from '../../data/sankey';
 import { createDiv, mount } from '../../utils/dom';
 
 describe('polygon', () => {
@@ -65,65 +64,6 @@ describe('polygon', () => {
         },
       },
       coordinate: [{ type: 'polar' }],
-    });
-
-    mount(createDiv(), chart);
-  });
-
-  it('render({...}) should render sankey plot', () => {
-    const chart = render<G2Spec>({
-      type: 'view',
-      data: SANKEY_DATA,
-      paddingBottom: 5,
-      paddingTop: 5,
-      paddingLeft: 5,
-      paddingRight: 5,
-      transform: [
-        {
-          type: 'sankey',
-        },
-      ],
-      scale: {
-        x: { guide: null },
-        y: { guide: null },
-        color: { guide: null },
-      },
-      children: [
-        {
-          type: 'polygon',
-          transform: [
-            {
-              type: 'connector',
-              callback: (v) => v.links,
-            },
-          ],
-          encode: {
-            x: 'x',
-            y: 'y',
-            color: (d) => d.source.name,
-            shape: 'ribbon',
-          },
-          style: {
-            fillOpacity: 0.5,
-            stroke: null,
-          },
-        },
-        {
-          type: 'polygon',
-          transform: [
-            {
-              type: 'connector',
-              callback: (v) => v.nodes,
-            },
-          ],
-          encode: {
-            x: 'x',
-            y: 'y',
-            color: 'name',
-            shape: 'polygon',
-          },
-        },
-      ],
     });
 
     mount(createDiv(), chart);
