+++ src/mark/geometry/link.ts
@@ -1,19 +1,29 @@
-import { MarkComponent as MC } from '../../runtime';
+import { MarkComponent as MC, Vector2 } from '../../runtime';
 import { LinkGeometry } from '../../spec';
 import {
   baseGeometryChannels,
   basePostInference,
   basePreInference,
 } from '../utils';
-import { Edge } from './edge';
 
 export type LinkOptions = Omit<LinkGeometry, 'type'>;
 
 /**
- * Convert value for each channel to link shapes. Same with edge.
+ * Connect `start` to `end` with single line.
  */
-export const Link: MC<LinkOptions> = (...args) => {
-  return Edge(...args);
+export const Link: MC<LinkOptions> = () => {
+  return (index, scale, value, coordinate) => {
+    const { x: X, y: Y, x1: X1, y1: Y1 } = value;
+
+    const xoffset = scale.x?.getBandWidth?.() || 0;
+
+    const P = index.map((i) => [
+      coordinate.map([+X[i] + xoffset / 2, +Y[i]]) as Vector2,
+      coordinate.map([+X1[i] + xoffset / 2, +Y1[i]]) as Vector2,
+    ]);
+
+    return [index, P];
+  };
 };
 
 Link.props = {
