+++ src/mark/geometry/polygon.ts
@@ -49,5 +49,5 @@ Polygon.props = {
     { type: 'maybeTitleX' },
     { type: 'maybeTooltipY' },
   ],
-  shapes: ['polygon', 'ribbon'],
+  shapes: ['polygon'],
 };
