+++ src/shape/edge/ribbon.ts
