+++ src/shape/index.ts
@@ -44,15 +44,14 @@ export { SmoothArea, SmoothAreaOptions } from './area/smoothArea';
 export { Link, LinkOptions } from './link/link';
 export { Image, ImageOptions } from './image/image';
 export { Polygon, PolygonOptions } from './polygon/polygon';
-export { Ribbon, RibbonOptions } from './polygon/ribbon';
 export { Box, BoxOptions } from './schema/box';
-
 // For relation diagram (node and edge).
 export { Point as PointNode } from './node/point';
 export { Polygon as PolygonNode } from './node/polygon';
 export { Edge, EdgeOptions } from './edge/edge';
 export { SmoothEdge, SmoothEdgeOptions } from './edge/smoothEdge';
 export { Arc, ArcOptions } from './edge/arc';
+export { Ribbon, RibbonOptions } from './edge/ribbon';
 
 export {
   AnnotationLine as AnnotationLine,
